moved Ui.java into its own package
