+++ src/main/java/dolla/command/InitialModifyCommand.java
@@ -1,7 +1,7 @@
 package dolla.command;
 
 import dolla.DollaData;
-import dolla.Ui;
+import dolla.ui.Ui;
 import dolla.task.LogList;
 
 public class InitialModifyCommand extends Command {
