+++ src/main/java/dolla/command/ShowListCommand.java
@@ -1,14 +1,9 @@
 package dolla.command;
 
 import dolla.DollaData;
-import dolla.task.Log;
-import dolla.Ui;
+import dolla.ui.Ui;
 
-import dolla.task.EntryList;
-import dolla.task.DebtList;
 import dolla.task.LogList;
-import dolla.task.Task;
-import dolla.task.TaskList;
 
 
 import java.util.ArrayList;
