+++ src/main/java/dolla/command/SnoozeCommand.java
@@ -1,7 +1,7 @@
 package dolla.command;
 
 import dolla.Time;
-import dolla.Ui;
+import dolla.ui.Ui;
 import dolla.task.Deadline;
 import dolla.task.Event;
 import dolla.task.Task;
