+++ src/main/java/dolla/command/SwitchModeCommand.java
@@ -1,7 +1,7 @@
 package dolla.command;
 
 import dolla.DollaData;
-import dolla.Ui;
+import dolla.ui.Ui;
 
 public class SwitchModeCommand extends Command {
     private String newMode;
