+++ src/main/java/dolla/parser/DollaParser.java
@@ -1,6 +1,6 @@
 package dolla.parser;
 
-import dolla.Ui;
+import dolla.ui.Ui;
 import dolla.command.AddDebtsCommand;
 import dolla.command.AddEntryCommand;
 import dolla.command.Command;
