+++ src/main/java/dolla/parser/EntryParser.java
@@ -1,6 +1,5 @@
 package dolla.parser;
 
-import dolla.Ui;
 import dolla.command.*;
 
 public class EntryParser extends Parser { public EntryParser(String inputLine) {
