+++ src/main/java/dolla/ui/LimitUi.java
@@ -1,6 +1,5 @@
 package dolla.ui;
 
-import dolla.Ui;
 import dolla.task.Limit;
 
 /**
