+++ src/main/java/dolla/ui/Ui.java
@@ -1,4 +1,4 @@
-package dolla;
+package dolla.ui;
 
 import dolla.task.*;
 
@@ -8,7 +8,7 @@ import java.util.Arrays;
 /**
  * duke.Ui is a class that handles all interactions to the user.
  */
-public class Ui {
+public abstract class Ui {
 
     private static String logo = " ____        _        \n"
             + "|  _ \\ _   _| | _____ \n"
@@ -16,8 +16,7 @@ public class Ui {
             + "| |_| | |_| |   <  __/\n"
             + "|____/ \\__,_|_|\\_\\___|\n";
 
-    //to be changed back to private after class is moved to dolla.ui package
-    public static String line = "\t____________________________________________________________";
+    private static String line = "\t____________________________________________________________";
 
     public static void showWelcome() {
         System.out.println("Hello from\n" + logo);
